tagging-recording/requirements.txt
cryptography==3.0
PyJWT==1.7.1
requests==2.24.0
